Refactor : [https://github.com/Web-Developers2-0/2025-Practica_Profesionalizante/issues/46], code refactoring and installation of new apps
