+++ backend/PlanetSuperheroes/settings.py
@@ -65,6 +65,10 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',
     'rest_framework_simplejwt.token_blacklist',
     'User',
+    'Product',
+    'Order',
+    'Event',
+    'Notification'
 ]
 
 MIDDLEWARE = [
@@ -82,7 +86,9 @@ MIDDLEWARE = [
 
 ]
 CORS_ORIGIN_WHITELIST = ['http://localhost:4200']
-CORS_ALLOWED_ORIGINS = ['http://localhost:4200',]
+CORS_ALLOWED_ORIGINS = [
+    "http://localhost:4200",  # Cambia esto por el puerto o dominio donde esté corriendo tu frontend
+]
 CORS_ALLOW_CREDENTIALS = True
 ROOT_URLCONF = 'PlanetSuperheroes.urls'
