+++ backend/PlanetSuperheroes/urls.py
@@ -43,7 +43,8 @@ schema_view = get_schema_view(
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('User.urls')),
-    
+    path('api/', include('Product.urls')),
+    path('api/', include('Order.urls')),
