+++ backend/Product/serializers.py
@@ -1,0 +1,12 @@
+from Product.models import Category, Product
+from rest_framework import serializers
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = '__all__'
+        
+class ProductSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Product
+        fields = '__all__'
