+++ backend/Product/urls.py
@@ -1,0 +1,13 @@
+# backend/Product/urls.py
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+from .views import CategoryViewSet, ProductViewSet
+
+# Crear un router y registrar las vistas
+router = DefaultRouter()
+router.register(r'categories', CategoryViewSet)
+router.register(r'products', ProductViewSet)
+
+urlpatterns = [
+    path('', include(router.urls)),  # Esto incluirá las rutas generadas automáticamente
+]
