+++ backend/User/models.py
@@ -3,6 +3,7 @@ from django.db import models
 from django.utils import timezone
 from django.contrib.auth.models import Group 
 
+
 class UserManager(BaseUserManager):
     def create_user(self, email, password=None, role=None, **extra_fields):
         if not email:
@@ -25,7 +26,6 @@ class UserManager(BaseUserManager):
 
 class User(AbstractBaseUser, PermissionsMixin):
     id = models.AutoField(primary_key=True) 
-    
     email = models.EmailField(unique=True)
     username = models.CharField(max_length=30, default='',blank=False, null=True) 
     first_name = models.CharField(max_length=30,default='', blank=False)
@@ -47,8 +47,11 @@ class User(AbstractBaseUser, PermissionsMixin):
 
     def __str__(self):
         return self.email
+    
     def user_orders(self):
+        from Order.models import Order  # import interno (lazy import)
         return Order.objects.filter(id_user=self)
+
     
     class Meta:
         db_table = 'user'
@@ -65,72 +68,3 @@ class Role(models.Model):
 
     def __str__(self):
         return self.name
-
-class Category(models.Model):
-    id_category = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=45, blank=False)
-    
-    class Meta:
-        db_table = 'categories'
-        verbose_name = 'Category'
-        verbose_name_plural = 'Categories'
-        
-    def __str__(self):
-        return self.name    
-
-
-class Product(models.Model):
-    id_product = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=100, blank=False)
-    description = models.CharField(max_length=5000, blank=False)
-    price = models.DecimalField(max_digits=10, decimal_places=2, blank=False)
-    discount = models.IntegerField(blank=True, null=True)
-    stock = models.IntegerField(blank=False)
-    image = models.CharField(max_length=255, blank=True, null=True)
-    pages = models.IntegerField(blank=True, null=True)
-    format = models.CharField(max_length=45, blank=True, null=True)
-    weight = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
-    isbn = models.CharField(max_length=45, blank=True, null=True)
-    category = models.ForeignKey(Category, on_delete=models.CASCADE)
-    calification = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
-    
-    class Meta: 
-        db_table = 'products'
-        verbose_name = 'Product'
-        verbose_name_plural = 'Products'
-        
-    def __str__(self):
-        return self.name
-
-class Order(models.Model):
-    id_order = models.AutoField(primary_key=True)
-    id_user = models.ForeignKey(User, on_delete=models.CASCADE, null=True, db_column='user_id', related_name='orders')
-    state = models.CharField(max_length=45, blank=True)
-    order_date = models.DateField(null=True)
-    payment_method = models.CharField(max_length=45, blank=True)
-    shipping_method = models.CharField(max_length=45, null=True)
-    payment_status = models.CharField(max_length=45, null=True)
-    total_amount = models.DecimalField(max_digits=10, decimal_places=2, null=True)
-
-    class Meta:
-        db_table = 'orders'
-        verbose_name = 'Order'
-        verbose_name_plural = 'Orders'
-
-    def __str__(self):
-        return f'Order {self.id_order}'
-
-
-class OrderItem(models.Model):
-    id_order_items = models.AutoField(primary_key=True)
-    quantity = models.IntegerField(blank=False)
-    product = models.ForeignKey(Product, on_delete=models.CASCADE, null=True, related_name='order_items')
-    order = models.ForeignKey(Order, on_delete=models.CASCADE)
-    
-    class Meta:
-        db_table = 'order_items'
-        verbose_name = 'Order Item'
-        verbose_name_plural = 'Order Items'
-        
-    def __str__(self):
-        return f'{self.quantity} of {self.product.name} in Order {self.order.id_order}'
