+++ backend/User/views.py
@@ -6,9 +6,8 @@ from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework.generics import GenericAPIView
 from rest_framework.generics import RetrieveUpdateAPIView
 from django.contrib.auth import authenticate
-from .serializers import RoleSerializer, UserSerializer, CustomTokenObtainPairSerializer, ProductSerializer, CategorySerializer,OrderCreateSerializer, OrderSerializer,LogoutSerializer
-from .models import Role, User, Product, Category, Order
-from rest_framework.generics import ListAPIView
+from .serializers import RoleSerializer, UserSerializer, CustomTokenObtainPairSerializer,LogoutSerializer
+from .models import Role, User
 from User.serializers import (CustomTokenObtainPairSerializer, UserSerializer)
 from User.models import User
 from rest_framework.permissions import AllowAny
@@ -17,8 +16,6 @@ from rest_framework.viewsets import ModelViewSet
 from django.utils import timezone
 from rest_framework.decorators import api_view
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser
-
-
 
 class RegisterView(APIView):
     permission_classes = [AllowAny]
@@ -29,7 +26,6 @@ class RegisterView(APIView):
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
 
 class Login(TokenObtainPairView):
     serializer_class = CustomTokenObtainPairSerializer
@@ -51,7 +47,6 @@ class Login(TokenObtainPairView):
                 }, status=status.HTTP_200_OK)
             return Response({'error': 'Contraseña o nombre de usuario incorrectos'}, status=status.HTTP_400_BAD_REQUEST)
         return Response({'error': 'Contraseña o nombre de usuario incorrectos'}, status=status.HTTP_400_BAD_REQUEST)
-
  
 class Logout(GenericAPIView):
     permission_classes = [IsAuthenticated]
@@ -62,11 +57,8 @@ class Logout(GenericAPIView):
             RefreshToken.for_user(user.first())
             return Response({'message': 'Sesión cerrada correctamente.'}, status=status.HTTP_200_OK)
         return Response({'error': 'No existe este usuario.'}, status=status.HTTP_400_BAD_REQUEST)
-         
-
-        
+                
 # ver los datos del usuario logueado
-
 class UserView(RetrieveUpdateAPIView):
     serializer_class = UserSerializer
     permission_classes = [IsAuthenticated]
@@ -74,7 +66,6 @@ class UserView(RetrieveUpdateAPIView):
     def get_object(self):
         return self.request.user
     
-
 # actualizar los datos del usuario 
 class UpdateUserView(APIView):
     serializer_class = UserSerializer
@@ -91,57 +82,7 @@ class UpdateUserView(APIView):
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
-class CategoryViewSet(ModelViewSet):
-    queryset = Category.objects.all()
-    serializer_class = CategorySerializer
-    permission_classes = [AllowAny]
-
-""" class ProductViewSet(ModelViewSet):
-    queryset = Product.objects.all()
-    serializer_class = ProductSerializer
-    permission_classes = [AllowAny] """
-    
-
-class ProductViewSet(ModelViewSet):
-    queryset = Product.objects.all()
-    serializer_class = ProductSerializer
-    permission_classes = [AllowAny]
-
-
-    def get_permissions(self):
-        if self.action in ['create', 'update', 'partial_update', 'destroy']:
-            self.permission_classes = [IsAdminUser]
- 
-        return super(ProductViewSet, self).get_permissions()
-    
-
-#CREAR ORDENES CON USUARIO AUTENTICADO 
-class CreateOrderView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request, *args, **kwargs):
-        serializer = OrderCreateSerializer(data=request.data, context={'request': request})
-        if serializer.is_valid():
-            order = serializer.save(id_user=request.user)
-            return Response(OrderSerializer(order).data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-
-
-    
-#VER LISTA DE ORDENES DE USUARIO AUTENTICADO      
-class UserOrdersView(ListAPIView):
-    permission_classes = [IsAuthenticated]
-    serializer_class = OrderSerializer
-
-    def get_queryset(self):
-        user_id = self.request.user.id
-        return Order.objects.filter(id_user=user_id)    
-
-
-
 class RoleViewSet(ModelViewSet):
     queryset = Role.objects.all()
     serializer_class = RoleSerializer
     permission_classes = [IsAdminUser]
-
